fix: graph client bug
