+++ apps/evm/package.json
@@ -60,6 +60,7 @@
     "framer-motion": "7.10.3",
     "lodash.once": "4.1.1",
     "lodash.zip": "4.2.0",
+    "ioredis": "5.3.2",
     "nanoid": "4.0.0",
     "next": "14.1.0",
     "next-axiom": "1.1.1",
