--- apps/evm/src/lib/bar/index.ts
@@ -1,2 +1,0 @@
-export * from './useBarChartData'
-export * from './useBarData'
