+++ apps/evm/src/lib/functions.ts
@@ -26,7 +26,7 @@ import {
 
 import { Bound } from './constants'
 import { useTicks } from './hooks'
-import { TickProcessed } from './hooks/useConcentratedActiveLiquidity'
+import { TickProcessed } from './pool/v3/use-concentrated-active-liquidity'
 
 export const isTridentConstantPool = (
   pool: SushiSwapV2Pool | TridentConstantPool | TridentStablePool | null,
