+++ apps/evm/src/lib/hooks/index.ts
@@ -1,8 +1,8 @@
 export * from './api'
-export * from './useIsTickAtLimit'
+export * from '../pool/v3/use-is-tick-at-limit'
 export * from './usePersistedSlippageError'
 export * from './usePriceInverter'
-export * from './useTicks'
+export * from '../pool/v3/use-ticks'
 export * from './useTokenAmountDollarValues'
 export * from './useTokensFromPool'
 export * from './useTransactionDeadline'
