+++ apps/evm/src/lib/pool/trident/actions/burnLiquidityAction.ts
