+++ apps/evm/src/lib/stake/use-bar-chart-data.ts
