+++ apps/evm/src/lib/stake/use-bar-data.ts
