+++ apps/evm/src/ui/analytics/furo-token-table.tsx
@@ -19,7 +19,7 @@ import {
   FuroToken,
   GetFuroTokenArgs,
   useFuroTokens,
-} from '../../lib/furo/useFuroTokens'
+} from '../../lib/analytics/use-furo-tokens'
 import { usePoolFilters } from '../pool'
 
 const COLUMNS: ColumnDef<FuroToken, unknown>[] = [
