+++ apps/evm/src/ui/swap/cross-chain/use-is-cross-chain-swap-maintenance.ts
@@ -1,12 +1,12 @@
 import { useQuery } from '@tanstack/react-query'
-import { CrossChainSwapEdgeConfig } from 'src/lib/edge/get-cross-chain-swap-edge-config'
+import { CrossChainSwapEdgeConfig } from 'src/app/swap/cross-chain/get-cross-chain-swap-edge-config'
 import { useEdgeConfig } from 'src/providers/edge-config-provider'
 
 export const useIsCrossChainSwapMaintenance = () => {
   const { maintenance } = useEdgeConfig<CrossChainSwapEdgeConfig>()
 
   return useQuery({
-    queryKey: ['useIsCrossChainSwapMaintenance'],
+    queryKey: ['cross-chain-swap-maintenance'],
     queryFn: async () => {
       const resp = await fetch('/api/config/xswap', {
         next: { revalidate: 60 },
