+++ packages/client/package.json
@@ -66,7 +66,6 @@
   },
   "dependencies": {
     "@sushiswap/database": "workspace:*",
-    "@sushiswap/token-price-api": "workspace:*",
     "@sushiswap/viem-config": "workspace:*",
     "@whatwg-node/fetch": "0.8.4",
     "deepmerge-ts": "^4.3.0",
