+++ packages/graph-client/resolvers/concentrated/v3factoriesByChainIds.ts
@@ -10,6 +10,7 @@ import {
   QueryResolvers,
   SUSHISWAP_V3_Factory,
 } from '../../.graphclient/index.js'
+import { isPromiseFulfilled } from 'sushi/validate'
 
 export const v3factoriesByChainIds: QueryResolvers['v3factoriesByChainIds'] =
   async (
@@ -18,36 +19,39 @@ export const v3factoriesByChainIds: QueryResolvers['v3factoriesByChainIds'] =
     context,
     info,
   ): Promise<Query['v3factoriesByChainIds']> => {
-    const fetchFactory = async (
-      chainId: SushiSwapV3ChainId,
-    ): Promise<SUSHISWAP_V3_Factory> => {
-      return context.SushiSwapV3.Query.SUSHISWAP_V3_factories({
-        root,
-        args,
-        context: {
-          ...context,
-          chainId,
-          subgraphName: SUSHISWAP_V3_SUBGRAPH_NAME[chainId],
-          subgraphHost: SUBGRAPH_HOST[chainId],
-        },
-        info,
-      }).then((factories: SUSHISWAP_V3_Factory[]) => ({
-        ...factories[0],
-        chainId,
-      }))
-    }
-
-    const factories = await Promise.allSettled(
+    return Promise.allSettled(
       args.chainIds
         .filter((chainId): chainId is SushiSwapV3ChainId =>
           SUSHISWAP_V3_ENABLED_NETWORKS.includes(chainId),
         )
-        .map((chainId) => fetchFactory(chainId)),
-    )
-
-    return factories
-      .map((settlement) =>
-        settlement.status === 'fulfilled' ? settlement.value : null,
-      )
-      .filter((factory): factory is SUSHISWAP_V3_Factory => factory !== null)
+        .map((chainId) => {
+          return context.SushiSwapV3.Query.SUSHISWAP_V3_factories({
+            root,
+            args,
+            context: {
+              ...context,
+              chainId,
+              subgraphName: SUSHISWAP_V3_SUBGRAPH_NAME[chainId],
+              subgraphHost: SUBGRAPH_HOST[chainId],
+            },
+            info,
+          }).then((factories: SUSHISWAP_V3_Factory[]) => {
+            return factories?.length > 0
+            ? factories.map((factory) => ({
+                ...factory,
+                chainId,
+              }))
+            : []
+          })
+        }),
+    ).then((promiseSettledResults) => {
+      if (!Array.isArray(promiseSettledResults)) {
+        console.error('v3 factories query failed...', promiseSettledResults)
+        return []
+      }
+      return promiseSettledResults
+        .flat()
+        .filter(isPromiseFulfilled)
+        .flatMap((promiseFulfilled) => promiseFulfilled.value)
+    })
   }
